@@ -32,27 +32,23 @@
     "filter"
   ],
   "homepage": "https://github.com/bithost-gmbh/ngx-mat-select-search",
-  "peerDependencies": {
-    "@angular/core": "^5.0.0 || ^6.0.0 || ^7.0.0 || ^8.0.0",
-    "@angular/material": "^5.0.0 || ^6.0.0 || ^7.0.0 || ^8.0.0",
-    "rxjs": "^5.5.2 || ^6.0.0"
-  },
+  "peerDependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.4",
-    "@angular/animations": "^7.2.0",
-    "@angular/cdk": "^7.2.1",
-    "@angular/cli": "^7.1.4",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/forms": "^7.2.0",
+    "@angular-devkit/build-angular": "^0.803.9",
+    "@angular/animations": "^8.2.10",
+    "@angular/cdk": "^8.2.3",
+    "@angular/cli": "^8.3.9",
+    "@angular/common": "^8.2.10",
+    "@angular/compiler": "^8.2.10",
+    "@angular/compiler-cli": "^8.2.10",
+    "@angular/core": "^8.2.10",
+    "@angular/forms": "^8.2.10",
     "@angular/http": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
+    "@angular/language-service": "^8.2.10",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.10",
+    "@angular/platform-browser-dynamic": "^8.2.10",
+    "@angular/router": "^8.2.10",
     "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "~10.12.18",
@@ -67,11 +63,11 @@
     "karma-jasmine-html-reporter": "^1.4.0",
     "ng-packagr": "^4.4.5",
     "protractor": "~5.4.2",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.3",
     "ts-node": "~7.0.1",
     "tslint": "~5.12.0",
     "tsickle": "^0.34.0",
-    "typescript": "~3.2.2",
-    "zone.js": "^0.8.19"
+    "typescript": "~3.5.3",
+    "zone.js": "^0.9.1"
   }
 }
