@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import {VERSION} from '@angular/material';
+import { VERSION } from '@angular/material/core';
 
 import { MatSelectSearchVersion } from './mat-select-search/ngx-mat-select-search.module';
 
